# docs/swagger.yaml
basePath: /api
definitions:
...
  structs.CancelTransaksiObatResponse:
    properties:
      data:
        $ref: '#/definitions/structs.TransaksiObat'
      pesan:
        example: Transaksi obat berhasil dibatalkan
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.DeleteTransaksiObatResponse:
    properties:
      data:
        $ref: '#/definitions/structs.TransaksiObat'
      pesan:
        example: Transaksi resep obat berhasil dihapus.
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.ErrorResponse:
    properties:
      pesan:
        example: Terjadi kesalahan
        type: string
      status:
        example: error
        type: string
    type: object
...
  structs.GetTransaksiObatByIDResponse:
    properties:
      data:
        $ref: '#/definitions/structs.TransaksiObat'
      pesan:
        example: Get transaksi obat by ID berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
  structs.GetTransaksiObatResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/structs.TransaksiObat'
        type: array
      pesan:
        example: Get transaksi obat berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.TambahTransaksiObatRequest:
    properties:
      pasien_id:
        type: integer
    required:
    - pasien_id
    type: object
  structs.TambahTransaksiObatResponse:
    properties:
      data:
        $ref: '#/definitions/structs.TransaksiObat'
      pesan:
        example: Transaksi obat berhasil ditambahkan
        type: string
      status:
        example: sukses
        type: string
    type: object
...
  structs.TransaksiObat:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      dibayar_at:
        type: string
      grand_total:
        type: number
      id:
        type: integer
      kembalian:
        type: number
      modified_at:
        type: string
      modified_by:
        type: string
      nama_pasien:
        type: string
      no_resep:
        type: string
      pasien_id:
        type: integer
      status:
        type: string
      tipe_pembayaran:
        type: string
      total_pembayaran:
        type: number
    type: object
...
  structs.UpdateTransaksiObatRequest:
    properties:
      status:
        enum:
        - PENDING
        - CANCELED
        type: string
    type: object
  structs.UpdateTransaksiObatResponse:
    properties:
      data:
        $ref: '#/definitions/structs.TransaksiObat'
      pesan:
        example: Update data transaksi resep obat pasien berhasil
        type: string
      status:
        example: sukses
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: '-'
    name: Damar
  description: 'Dokumentasi REST API Alur Obat Farmasi menggunakan Golang (Framework
    : Gin dan Database : PostgreSQL).'
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Final Project API
  version: "1.0"
paths:
...
  /api/admin-farmasi/transaksi-obat:
    get:
      description: Menampilkan daftar transaksi obat
      produces:
      - application/json
      responses:
        "200":
          description: Berhasil Get daftar transaksi resep obat
          schema:
            $ref: '#/definitions/structs.GetTransaksiObatResponse'
        "400":
          description: Gagal Get data transaksi resep obat
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Transaksi resep obat tidak ada
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Semua Transaksi Obat
      tags:
      - 6. Transaksi Obat
    post:
      consumes:
      - application/json
      description: Membuat transaksi obat baru
      parameters:
      - description: Data Transaksi Obat
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.TambahTransaksiObatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Tambah data transaksi resep obat pasien berhasil
          schema:
            $ref: '#/definitions/structs.TambahTransaksiObatResponse'
        "400":
          description: Gagal tambah data transaksi resep obat
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: Transaksi resep obat sudah dibuat
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tambah Transaksi Obat
      tags:
      - 6. Transaksi Obat
  /api/admin-farmasi/transaksi-obat/{id}:
    delete:
      description: Menghapus transaksi obat
      parameters:
      - description: ID Transaksi Obat
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Transaksi resep obat berhasil dihapus
          schema:
            $ref: '#/definitions/structs.DeleteTransaksiObatResponse'
        "400":
          description: Gagal hapus transaksi resep obat
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: ID Transaksi resep obat tidak ditemukan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete Transaksi Obat
      tags:
      - 6. Transaksi Obat
    get:
      description: Menampilkan detail transaksi obat
      parameters:
      - description: ID Transaksi
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Get transaksi resep obat by ID berhasil
          schema:
            $ref: '#/definitions/structs.GetTransaksiObatByIDResponse'
        "400":
          description: Gagal Get data transaksi resep obat by ID
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: ID  Transaksi resep obat tidak ditemukan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Transaksi Obat By ID
      tags:
      - 6. Transaksi Obat
    put:
      consumes:
      - application/json
      description: Mengubah data transaksi obat
      parameters:
      - description: ID Transaksi Obat
        in: path
        name: id
        required: true
        type: integer
      - description: Data Transaksi resep obat
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/structs.UpdateTransaksiObatRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Update data transaksi resep obat pasien berhasil
          schema:
            $ref: '#/definitions/structs.UpdateTransaksiObatResponse'
        "400":
          description: Gagal update transaksi resep obat
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: ID Transaksi resep obat tidak ditemukan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update Transaksi Obat
      tags:
      - 6. Transaksi Obat
  /api/admin-farmasi/transaksi-obat/{id}/cancel:
    patch:
      description: Membatalkan transaksi obat dengan mengembalikan stok obat jika
        status transaksi masih PENDING
      parameters:
      - description: ID Transaksi Obat
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Cancel data transaksi resep obat pasien berhasil
          schema:
            $ref: '#/definitions/structs.CancelTransaksiObatResponse'
        "400":
          description: Gagal hapus transaksi resep obat
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "401":
          description: ID Transaksi resep obat tidak ditemukan
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/structs.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cancel Transaksi Obat
      tags:
      - 6. Transaksi Obat
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...